+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New UI script runner feature. Launches _bellplay~_ if needed and loads script via OSC.
 - GUI: currently loaded script tab now displays emoji.
+- Pop-up docs: Add support for argument description, including those with options.
 
 ### Fixed